improve structure
