+++ tests/utils/common.test.js
@@ -1,6 +1,6 @@
 const chai = require('chai');
 const chaiAsPromised = require('chai-as-promised');
-const common = require('../../src/utils/common');
+const common = require('../../src/common/constants');
 
 chai.use(chaiAsPromised);
 const {
@@ -12,9 +12,6 @@ describe('test for common file', () => {
   it('should return common as object', () => {
     expect(common).to.not.be.undefined;
     expect(common).to.be.an('object');
-  });
-  it('should common has md5 as method', () => {
-    expect(typeof(common.md5)).to.eql('function');
   });
   it('should common has properties', () => {
     expect(typeof(common.FACEBOOK)).to.eql('string');
@@ -36,10 +33,4 @@ describe('test for common file', () => {
     expect((common.YOUTUBE)).to.eql('youtube');
     expect((common.TTL_REDIS)).to.eql(3600);
   });
-  describe('test md5 method', () => {
-    it('should return correct output', () => {
-      expect(common.md5('botsaris.d@gmail.com')).to.eql('db35cedc94bc259282e658234a256c59');
-    });
-  });
-
 });
